chore: rework file hierarchy
