+++ src/main/java/school/hei/haapi/endpoint/model/StudentResource.java
@@ -1,0 +1,18 @@
+package school.hei.haapi.endpoint.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class StudentResource {
+  String id;
+  String firstName;
+  String lastName;
+  String email;
+  String role;
+}
