+++ src/main/java/school/hei/haapi/endpoint/rest/controller/HealthController.java
@@ -1,0 +1,13 @@
+package school.hei.haapi.endpoint.rest.controller;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class HealthController {
+
+  @GetMapping("/ping")
+  public String hello() {
+    return "pong";
+  }
+}
