+++ src/main/java/school/hei/haapi/endpoint/rest/security/model/ApiClient.java
@@ -1,0 +1,62 @@
+package school.hei.haapi.endpoint.rest.security.model;
+
+import java.util.Collection;
+import java.util.List;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.ToString;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import school.hei.haapi.model.User;
+
+
+@Getter
+@AllArgsConstructor
+@ToString
+public class ApiClient implements UserDetails {
+
+  private final User user;
+
+  @Override
+  public Collection<? extends GrantedAuthority> getAuthorities() {
+    return List.of(Role.valueOf(user.getRole()));
+  }
+
+  @Override
+  public String getPassword() {
+    return null;
+  }
+
+  @Override
+  public String getUsername() {
+    return user.getLastName();
+  }
+
+  @Override
+  public boolean isAccountNonExpired() {
+    return true;
+  }
+
+  @Override
+  public boolean isAccountNonLocked() {
+    return true;
+  }
+
+  @Override
+  public boolean isCredentialsNonExpired() {
+    return true;
+  }
+
+  @Override
+  public boolean isEnabled() {
+    return true;
+  }
+
+  public String getUserId() {
+    return user.getId();
+  }
+
+  public String getRole() {
+    return user.getRole();
+  }
+}
