--- src/main/java/school/hei/haapi/exception/BadRequestException.java
@@ -1,7 +1,0 @@
-package school.hei.haapi.exception;
-
-public class BadRequestException extends ApiException {
-  public BadRequestException(String message) {
-    super(message, ExceptionType.CLIENT_EXCEPTION);
-  }
-}
