--- src/main/java/school/hei/haapi/exception/ForbiddenException.java
@@ -1,7 +1,0 @@
-package school.hei.haapi.exception;
-
-public class ForbiddenException extends ApiException {
-  public ForbiddenException(String message) {
-    super(message, ExceptionType.CLIENT_EXCEPTION);
-  }
-}
