--- src/main/java/school/hei/haapi/exception/NotFoundException.java
@@ -1,7 +1,0 @@
-package school.hei.haapi.exception;
-
-public class NotFoundException extends ApiException {
-  public NotFoundException(String message) {
-    super(message, ExceptionType.CLIENT_EXCEPTION);
-  }
-}
