--- src/main/java/school/hei/haapi/exception/NotImplementedException.java
@@ -1,7 +1,0 @@
-package school.hei.haapi.exception;
-
-public class NotImplementedException extends ApiException {
-  public NotImplementedException(String message) {
-    super(message, ExceptionType.SERVER_EXCEPTION);
-  }
-}
