--- src/main/java/school/hei/haapi/exception/TooManyRequestsException.java
@@ -1,11 +1,0 @@
-package school.hei.haapi.exception;
-
-public class TooManyRequestsException extends ApiException {
-  public TooManyRequestsException(String message) {
-    super(message, ExceptionType.CLIENT_EXCEPTION);
-  }
-
-  public TooManyRequestsException(Exception source) {
-    super(source, ExceptionType.CLIENT_EXCEPTION);
-  }
-}
