+++ src/main/java/school/hei/haapi/model/exception/BadRequestException.java
@@ -1,0 +1,7 @@
+package school.hei.haapi.model.exception;
+
+public class BadRequestException extends ApiException {
+  public BadRequestException(String message) {
+    super(message, ExceptionType.CLIENT_EXCEPTION);
+  }
+}
