+++ src/main/java/school/hei/haapi/model/exception/ForbiddenException.java
@@ -1,0 +1,7 @@
+package school.hei.haapi.model.exception;
+
+public class ForbiddenException extends ApiException {
+  public ForbiddenException(String message) {
+    super(message, ExceptionType.CLIENT_EXCEPTION);
+  }
+}
