+++ src/main/java/school/hei/haapi/model/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package school.hei.haapi.model.exception;
+
+public class NotFoundException extends ApiException {
+  public NotFoundException(String message) {
+    super(message, ExceptionType.CLIENT_EXCEPTION);
+  }
+}
