+++ src/main/java/school/hei/haapi/model/exception/NotImplementedException.java
@@ -1,0 +1,7 @@
+package school.hei.haapi.model.exception;
+
+public class NotImplementedException extends ApiException {
+  public NotImplementedException(String message) {
+    super(message, ExceptionType.SERVER_EXCEPTION);
+  }
+}
