--- src/main/java/school/hei/haapi/security/model/Role.java
@@ -1,18 +1,0 @@
-package school.hei.haapi.security.model;
-
-import org.springframework.security.core.GrantedAuthority;
-
-public enum Role implements GrantedAuthority {
-  STUDENT,
-  TEACHER,
-  MANAGER;
-
-  public String getRole() {
-    return name();
-  }
-
-  @Override
-  public String getAuthority() {
-    return "ROLE_" + getRole();
-  }
-}
