--- src/main/java/school/hei/haapi/web/model/TeacherResource.java
@@ -1,18 +1,0 @@
-package school.hei.haapi.web.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class TeacherResource {
-  String id;
-  String firstName;
-  String lastName;
-  String email;
-  String role;
-}
