--- src/main/java/school/hei/haapi/web/rest/HealthController.java
@@ -1,13 +1,0 @@
-package school.hei.haapi.web.rest;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HealthController {
-
-  @GetMapping("/ping")
-  public String hello() {
-    return "pong";
-  }
-}
